Histogram 3
Implemented the mechanism for positioning and spacing out bars, and
resizing bars on the count axis based on the number of matching nodes.

One notable issue is that with discrete variables with a count less than
the default tick count (10), the number of bars changes to match the
number of unique values but the ticks do not move.