NodeFile Edge Generation with PKey mapping
Implemented a potential performance boost that kicks in when a
non-linking file has a foreign key that maps directly to a file's
primary key, and fixed a bug in the nearby code that would have created
extra edges for every keypair in the Foreign Key.
Also added better documentation for the speed differences of the two
methods.