FilePicker Keyboard Nav Improvements
The file viewer now scrolls on selection change.
Selection will not jump to the last index if you press down from the
last row, nor will it jump to the first index if you press up in the
first row.
Moved the view out into the middle of the screen, where it looks much
better. It still needs to be made into a true modal dialog.