Window
Implemented the Window class, which provides a way of encapsulating GUI
functions in a draggable box. Unlike the Unity version, the value
returned by the function inside the Window will be passed through,
allowing for minimal changes to other code.

Also implemented functions for finding the left and right side of the
open screen in MenuController, similar to top and bottom.