Workspace Refactor 2
More work on serialization: Added the Workspace State concept which will
allow for serialization of any static variables, added DataFile
serialization (important!) and made DataFile.attributes private.

Also, Gridlines in graphing now default to off, because they cause a
significant reduction in frame rate when enabled. Need to come up with a
real fix for this.