Minor FilePicker improvements
The header bar for windows is now a button.
There is a title label indicating what the file picker is doing (set per
render)
Changing the functions on the FilePicker now refreshes the files inside,
which re-animates the icons to show that something is actually changing
when someone launches a File-picking action while they're already using
it.