GuiPlus depth 1
Implemented a basic method for forcing the GUI to treat depth as draw
order. All GUI calls rendered invisible and stored to be redrawn again
in multiple passes over the GUI depths that are used in the given frame.

Current implementation is more proof-of-concept and still contains some
ugly hacks to get working, and elements do not have their styles
respected.
Refactoring for next commit.