Fonts and Bugs
Added a custom font style that can be controlled by the from the
MainMenu, chose a new font for the program, and adjusted most of the
menus that were misaligned with the new font.

Bug fix / small refactor of FkeyMenu. It was previously possible to add
key pairs to a foreign key that didn't have a to-file specified, and
threw an exception when it started selecting the to-file attribute.

Fixed the old bug where toggling AttributeMenu from the FileMenu
wouldn't let you re-open the same attribute after closing it.