FilePicker List View
Added another style for viewing files which displays the icons as
one-per-row with a larger horizontal area for displaying the file name.
It's a lot easier to use so I made it the default. Added a toggle at the
bottom of the window to swap between the modes.
Also removed the foreign key editing from AttributeMenu. It's not worth
maintaining that code in two different places, especially when it
doesn't really add any value.