Fixed ScrollView mousePosition calculation
The GuiPlus.ScrollView call that automatically computes the size of the
inner content was not properly adjusting the scroll position on the
scroll pane stack. This is now fixed.

Also switched the FKey menu to use this version of the method and
changed the word "Foreign Keys" to "Edges" in several places