Auto-sizing ScrollView Fix
That fix from two commits ago was not the correct way to fix it, as it
broke the main functionality of the auto-sizing Scroll Views
(calculating the size). The bug was actually caused by the fact that a
auto-sizing Scroll View would never update it's position once in the
dictionary. Since the FKey menu slides from the left, it is initialized
with a position 0,0, which led to the calculateMousePosition() function
to use the un-updated position. Updating the position in BeginScrollView
fixes the problem.