Fixed object composition error
Unity 4.5 introduced an error message exposing a bug in its
serialization code, which processed private member variables on classes
(despite claims that it does NOT do that in the documentation). The
solution was to tag all private member variables with an attribute to
explicitly prevent serialization.
I also removed the dependency of Stats on Attribute.