Node Label Updates
Node labels are no longer visible beyond a certain distance. This cuts
about 20% of the CPU time spent per frame for positioning and rendering
unreadable labels, and almost doubles the framerate when fully zoomed
out.

Labels also have a higher minimum size, so if they are visible they are
easier to read.